Removed invalid test cases. Added Assertj tests
