+++ library/src/main/java/net/pubnative/library/models/PubnativeBeacon.java
@@ -2,12 +2,8 @@ package net.pubnative.library.models;
 
 public class PubnativeBeacon {
 
-    /**
-     * These are some Default beacons. We will add more in future.
-     */
-    public interface DefaultBeaconTypes{
-        String IMPRESSION = "impression";
-    }
+
+    public static final String IMPRESSION = "impression";
     public String type;
     public String url;
 }
